+++ .github/CHANGELOG.md
@@ -1,5 +1,6 @@
 # Changes from v1.2.1 to v
 1. [Game] Test - Added tests to check MilestoneAchievement, SustainedAchievement, AchievementManager classes
+2. [Game] Enhancement - Added an income from students
 
 # Changes from v1.2.0 to v1.2.1
 1. [Global] Bug - Remove reference to snapshot in stable release